bnx2x: add missing enum channel_tlvs definitions
Patchwork-id: 69648
O-Subject: [RHEL6.5 PATCH 4/4] bnx2x: add missing enum channel_tlvs definitions
Bugzilla: 1015137
RH-Acked-by: Jesper Brouer <jbrouer@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1015137

This fixes the interface that SRIOV virtual functions use to communicate with
the physical function. Since virtual functions can be assigned to virtual
guests running various operating systems, it is important that the interface is
the same for all of them and stable. We wouldn't be able to fix it in an
update, because that would break compatibility. We need to get the interface
right from the start.

Explanation from Yuval Mintz:
  The VF-PF channel is the infrastructure by which the VFs and a PF can
  communicate. Since they're sending TLVs to each other, they must agree upon
  the meaning of the TLV identifier, i.e., there mustn't be a PF and VF which
  understand a given identifer as different messages.

  To support this, for backward compatability the 'enum channel_tlvs' which
  defines the various messages can never have elements removed from it
  (although elements could get deprecated over time), and new TLVs should
  always be added at the end of the enum.

  While working on the VF RSS we've discovered that the upstream bnx2x version
  lacks a couple of entries in that enum (which were already defined in our
  non-upstream driver); those missing values were added by said upstream patch.

  The patch I've supplied takes only the neccessary changes to the enum so the
  the inbox RH6.5 driver will be compatible with other distros/drivers.
  Although there won't be a problem running a RH6.5 VM over a RH6.5 hypervisor
  (as both agree on the enum values) using current driver, there will be
  a major compatability issue with EVERY other driver version (including future
  RHEL products).

Upstream status:
These definitions come from upstream patch b9871bcf "bnx2x: VF RSS
support - PF side". We're not taking the whole patch at this moment because it
adds a new feature and a couple of known regressions.

Testing:
I tested that assignment of a VF to a virtual guest works. I could bring
the interface up in the guest, send/receive traffic and bring the
interface down successfully. It was necessary to have the patch applied
both in the host and in the guest. When the patch was applied only in the
host, bringing the interface down in the guest at the end of the test
triggered problems in the PF<->VF communication and timeouts, all due
to the PF<->VF interface mismatch. This was expected.

Signed-off-by: Rafael Aquini <aquini@redhat.com>