ext4: Remove warning from ext4_da_update_reserve_space()
Patchwork-id: 69819
O-Subject: [RHEL6.5 PATCH] ext4: Remove warning from ext4_da_update_reserve_space()
Bugzilla: 1011876
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

BZ 1011876
https://bugzilla.redhat.com/show_bug.cgi?id=1011876

BREW 6457837
https://brewweb.devel.redhat.com/taskinfo?taskID=6457837

This commit does not have upstream counterpart.

In commit ad4846f4a78dedc91174e582a94bb949a168064a we have introduced
warning in the case that number of allocated metadata block exceed number
of reserved metadata blocks.

However this might have been mistake, because even with rest of the
fixes we have in place it is still possible to happen in some
scenarios. We could backport more fixes which would narrow the
possibility of this happening, however those are complicated and RHEL6
codebase is very much different from upstream that it would have to be
basically implemented from scratch.

Luckily we have commit 86ad080b42971318389080a5f47b7859dce8565b which
introduces reserved space for exactly the purposes of underestimating
number of metadata blocks needed. So even in the unfortunate situation
of ENOSPC coming on wrong time we would still have a pool of blocks to
allocate metadata from, so it's very unlikely for this situation to
cause problems.

This commit silences the warning, however we introduce trace point
instead for the case that we need to investigate this in future.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>