ipv6: restrict neighbor entry creation to output flow
Patchwork-id: 69826
O-Subject: [patch rhel6.5 net] ipv6: restrict neighbor entry creation to output flow
Bugzilla: 997103
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Flavio Leitner <fbl@redhat.com>
RH-Acked-by: Francesco Fusco <ffusco@redhat.com>
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Daniel Borkmann <dborkman@redhat.com>

BZ997103
https://brewweb.devel.redhat.com/taskinfo?taskID=6461925

Upstream: http://patchwork.ozlabs.org/patch/281486/
this is queued up for stable trees since this is not applicable to
upstream tip. DaveM said he is going to push this to stable maintainers
very soon.

Description:
Commit 0d6a77079c475033cb622c07c5a880b392ef664e "ipv6: do not create
neighbor entries for local delivery" introduced a regression on
which routes to local delivery would not work anymore. Like this:

    $ ip -6 route add local 2001::/64 dev lo
    $ ping6 -c1 2001::9
    PING 2001::9(2001::9) 56 data bytes
    ping: sendmsg: Invalid argument

As this is a local delivery, that commit would not allow the creation of a
neighbor entry and thus the packet cannot be sent.

But as TPROXY scenario actually needs to avoid the neighbor entry creation only
for input flow, this patch now limits previous patch to input flow, keeping
output as before that patch.

Signed-off-by: Jiri Pirko <jpirko@redhat.com>
Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>