async: Revert MAX_THREADS to 256
Patchwork-id: 69853
O-Subject: [RHEL6.5 PATCH] async: Revert MAX_THREADS to 256
Bugzilla: 1021705
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

I just got another report in bz 1021705 that a previously working system is
hanging on boot.  Its happening because the system (a 4 core hp machine), has a
USB card reader on it that exports itself as about 8 scsi HBA's.  Because of
that, each hba preforms a scsi_host_scan using the async api, blocks on a
scsi_flush work, and at least one of those flushes blocks on an async
synchronization.  At this point there are unclaimed scsi host scans on the async
queue, and no available threads to service them, so we get deadlock.

QA has tested this patch which reverts the MAX_THREADS count to 256, like it was
before.  I'm concerned about a regression with bz 871955, but hopefully the
reschedule check will be enough to avoid that.  Regardless, at this point, I
give: 256 threads is the previous ceiling for async threads, and its pretty
arbitary (it seems like you should be able to deadlock with this value on any
system that has 257 or more HBA's), but thats a pretty rare system, and at this
late in the devel cycle I'd rather go back to what we know works for all our
available test systems, than try to figure a way around this

BZ: 1021705
Testing: Tested by QA, and indicated to work (see comment 17 in the bz)
Upstream: N/A
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6464867
Signed-off-by: Rafael Aquini <aquini@redhat.com>