nfsd: add private md5 hashing implementation to get around FIPS blacklisting
Patchwork-id: 69859
O-Subject: [RHEL6.6 PATCH] [BZ#1021076] nfsd: add private md5 hashing implementation to get around FIPS blacklisting
Bugzilla: 1020176
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

There are some known corner-cases in the NFSv4 spec that require the
server to track some information on stable storage in order to guarantee
that clients can properly recover state across a server reboot.

In order to do this, nfsd uses md5 to hash a large (up to 1k) opaque
blob down to a string that we can use as a dentry name. It uses the
kernel crypto API's md5 routines for this, but those get blacklisted
when the machine is booted with fips=1.

The choice to use a md5 hash here is unfortunate due to the
blacklisting, but since these hashes are not exposed anywhere outside of
the server there's no real reason to disallow them here.

This patch adds a "private" md5.c implementation for nfsd and fixes it
to use this code instead of the standard crypto APIs. I basically did
this by copying crypto/md5.c to fs/nfsd/md5.c, and then tweaked it to
strip out all of the crypto API goop.

This patch is not upstream. There, we added a new upcall program that
tracks this in an entirely different way. Backporting that would cause
grief for existing setups, so I think this is the best option for RHEL6.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>