nfs: Fix handling of partially delegated locks
Patchwork-id: 69862
O-Subject: [RHEL6.6 PATCH 2/2] [BZ#959788] NFSv4.x: Fix handling of partially delegated locks
Bugzilla: 959788
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

If a NFS client receives a delegation for a file after it has taken
a lock on that file, we can currently end up in a situation where
we mistakenly skip unlocking that file.

The following patch swaps an erroneous check in nfs4_proc_unlck for
whether or not the file has a delegation to one which checks whether
or not we hold a lock stateid for that file.

Reported-by: Chuck Lever <Chuck.Lever@oracle.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Cc: stable@vger.kernel.org [>=3.7]
Tested-by: Chuck Lever <Chuck.Lever@oracle.com>
Upstream commit: c5a2a15f8146fdfe45078df7873a6dc1006b3869

Signed-off-by: Rafael Aquini <aquini@redhat.com>