xfs: Ensure sync updates the log tail correctly
Patchwork-id: 70089
O-Subject: [RHEL 6.6 PATCH 2/2] xfs: Ensure sync updates the log tail correctly
Bugzilla: 1025439
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

From: David Chinner <dchinner@redhat.com>

From: David Chinner <dchinner@redhat.com>

RH BZ: 1025439
Upstream Commit: N/A, RHEL 6.5 regression fix

As a result of the change made in bb7aa47 ("[fs] xfs: prevent
deadlock trying to cover an active log"), the decisions to cover the
log made by xfs_log_need_covered() have changed. Because of the
difference between RHEL6 and upstream, it was not noticed that this
function is also called by xfs_quiesce_data() in the .sync_fs path
to ensure that the log tail is correctly updated as part of sync.
Hence the change of behaviour has caused a regression in xfs/182
where the log is not correctly recovered after a sync(1) call and
files are lost.

Fix it by always writing a dummy log record to the journal at the
end of xfs_quiesce_data to ensure that the tail of the log is always
up to date when sync completes.

Signed-off-by: David Chinner <dchinner@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>