revert xfs: prevent deadlock trying to cover an active log
Patchwork-id: 70461
O-Subject: [PATCH RHEL6.5] revert xfs: prevent deadlock trying to cover an active log
Bugzilla: 1014867

The fix for this bug, which was a deadlock:

[Bug 1014867] xfssyncd and flush device threads hang in xlog_grant_head_wait

caused a regression in integrity syncs; arguably worse.  And RHEL6.5 time
is extremely short.

So revert for now, to be followed quickly by a re-introduction, and 2 other
patches already on the list to fix the regression in RHEL6.6 and 6.5.Z.

Thanks,

-Eric