alt-smp-once should be alt-smp-boot
Patchwork-id: 70753
O-Subject: [RHEL6.6 PATCH BZ 1027505] Documentation: alt-smp-once should be alt-smp-boot
Bugzilla: 1027505
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1027505

RHEL6 only.

Upstream has moved on and no longer has this kernel parameter so this is
RHEL6 only.  alt-smp-once is not the right name for this parameter.  It is
alt-smp-boot in the code.

Signed-off-by: Rafael Aquini <aquini@redhat.com>