nfs: don't invalidate dentries that act as mountpoints
Patchwork-id: 70778
O-Subject: [RHEL6.6 PATCH] [BZ#1023276] nfs: don't invalidate dentries that act as mountpoints
Bugzilla: 1023276
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

We had a couple of reports of people that are mounting NFS filesystems,
and then bind mounting certain local files onto dentries in that nfs
mount (sort of like a poor-man's unionmount).

This all works well until the dentry serving as the mountpoint fails
d_revalidate. The dentry will end up being invalidated which makes the
bind mount unreachable via pathwalk.

Currently, nfs_revalidate_lookup will avoid invalidating directory
dentries when it or its children are mountpoints. It doesn't do any
such checking however for non-directories even though they can be
mountpoints too.

Fix nfs_revalidate_lookup to check d_mountpoint on any positive dentry
before invalidating it. If that returns true, then we need to consider
the dentry to be valid regardless of how the on-the-wire revalidation
went.

Note that this patch is not upstream. The problem was fixed there by
a different patchset that added check_submounts_and_drop(). That's a
bit much to backport for RHEL6 however.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>