fuse: never d_drop() dentry mountpoints on failed reval
Patchwork-id: 70801
O-Subject: [RHEL6.6 PATCH] fuse: never d_drop() dentry mountpoints on failed reval
Bugzilla: 1026952
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1026952

Upstream Status: N/A (see commit log for justification).

Build Info: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6556708

A non-directory can serve as a mountpoint for a bind mount. If a
non-dir serves as a bind mount target on a fuse fs and a
revalidate fails (e.g., due to replacement of the backing inode),
we d_drop() the dentry and the mountpoint is lost.

Extend the existing directory dentry mountpoint checks to
non-directories and refuse to drop the dentry if a mountpoint
exists.

There is no corresponding upstream commit for this change. Upstream
resolves this problem via the check_submounts_and_drop() helper,
which is too invasive for rhel6 and effectively backported to rhel6
fuse as of the following commit:

  c878e81c fuse: drop dentry on failed revalidate

Signed-off-by: Brian Foster <bfoster@redhat.com>

Tested: via inode replacement of a regular file used as a bind mount target on
  a glusterfs native client. Regression tested via the POSIX fs test suite and
  compile testing on a glusterfs volume.

Signed-off-by: Rafael Aquini <aquini@redhat.com>