ixgbe: fix inconsistent clearing of the multicast table
Patchwork-id: 70804
O-Subject: [PATCH RHEL6.5] ixgbe: fix inconsistent clearing of the multicast table
Bugzilla: 975248
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

It was reported that IPv6 hardware filtering tables were not being
correctly rewritten when using a bridge device with the PF.  This turned
out to be a driver error and was corrected after we notified Intel about
it.  This fix is a direct backport of upstream commit:

commit cf78959c0d7afbde31498afc4212294c28e2c278
Author: Emil Tantilov <emil.s.tantilov@intel.com>
Date:   Sat Oct 26 08:13:20 2013 +0000

    ixgbe: fix inconsistent clearing of the multicast table

    This patch resolves an issue where the MTA table can be cleared when the
    interface is reset while in promisc mode. As result IPv6 traffic between
    VFs will be interrupted.

    This patch makes the update of the MTA table unconditional to avoid the
    inconsistent clearing on reset.

Reporter verified that this will resolve RHBZ 975248.

Signed-off-by: Rafael Aquini <aquini@redhat.com>