cifs: stop trying to use virtual circuits
Patchwork-id: 70888
O-Subject: [RHEL6.6 PATCH] [bz 1026346]cifs: stop trying to use virtual circuits
Bugzilla: 1026346
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Fixes bz 1026346.

This is a backport of a patch sent recently committed into the upstream
kernel. The thread containing the patch is available at
http://thread.gmane.org/gmane.linux.kernel.cifs/8849

samba bz
https://bugzilla.samba.org/show_bug.cgi?id=10113

Copying over the commit message from the upstream post.
--
Currently, we try to ensure that we use vcnum of 0 on the first
established session on a connection and then try to use a different
vcnum on each session after that.

This is a little odd, since there's no real reason to use a different
vcnum for each SMB session. I can only assume there was some confusion
between SMB sessions and VCs. That's somewhat understandable since they
both get created during SESSION_SETUP, but the documentation indicates
that they are really orthogonal. The comment on max_vcs in particular
looks quite misguided. An SMB session is already uniquely identified
by the SMB UID value -- there's no need to again uniquely ID with a
VC.

Furthermore, a vcnum of 0 is a cue to the server that it should release
any resources that were previously held by the client. This sounds like
a good thing, until you consider that:

a) it totally ignores the fact that other programs on the box (e.g.
smbclient) might have connections established to the server. Using a
vcnum of 0 causes them to get kicked off.

b) it causes problems with NAT. If several clients are connected to the
same server via the same NAT'ed address, whenever one connects to the
server it kicks off all the others, which then reconnect and kick off
the first one...ad nauseum.

I don't see any reason to ignore the advice in "Implementing CIFS" which
has a comprehensive treatment of virtual circuits. In there, it states
"...and contrary to the specs the client should always use a VcNumber of
one, never zero."

Have the client just use a hardcoded vcnum of 1, and stop abusing the
special behavior of vcnum 0.

Reported-by: Sauron99@gmx.de <sauron99@gmx.de>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Volker Lendecke <vl@samba.org>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 9ae6cf606a33b0a762798df0fb742848bcc685b5)
--

This fixes a customer issue originally reported on RHEL 5. The customer
attempts to background mount multiple mounts from the same server which
due to an upstream bug can result in multiple tcp connections to the
same server. The problem arises when the server resets a session for
some reason. The client then reconnect the tcp session with the VC set
to 0 which results in the server resetting the other connections on the
client too. When the other clients attempt to reconnect, it ends up
resetting the first connection again. This results in a unending loop.

Patch has been built and simple sanity tests run on RHEL 6 machine. Main
customer testing was done on the RHEL 5 version of this patch.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>