gfs2: Add allocation parameters structure
Patchwork-id: 71000
O-Subject: [RHEL6.6 PATCH 1/4] GFS2: Add allocation parameters structure
Bugzilla: 998625
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

This patch adds a structure to contain allocation parameters with
the intention of future expansion of this structure. The idea is
that we should be able to add more information about the allocation
in the future in order to allow the allocator to make a better job
of placing the requests on-disk.

There is no functional difference from applying this patch.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>

rhbz#998625

Signed-off-by: Rafael Aquini <aquini@redhat.com>