evsel: Convert to _is_group_leader method
Patchwork-id: 71310
O-Subject: [PATCH RHEL6.6 BZ1034843 013/172] perf evsel: Convert to _is_group_leader method
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 823254edc66eb44bf612b1dfa4829afa9840f691
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Thu Nov 29 15:38:30 2012 +0900

description
===========
Convert perf_evsel__is_group_member to perf_evsel__is_group_leader.
This is because the most usecases are using negative form to check
whether the given evsel is a leader or not and it's IMHO somewhat
ambiguous - leader also *is* a member of the group.