evsel: Introduce method to request IDs be used
Patchwork-id: 71321
O-Subject: [PATCH RHEL6.6 BZ1034843 027/172] perf evsel: Introduce method to request IDs be used
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 7a5a5ca5fe3df8636c96b49591c23baf7c415dd4
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Dec 10 15:21:30 2012 -0300

description
===========
When mmaping multiple events we need to find the right evsel that
matches an event in the ring buffer.

For that we need to set the PERF_FORMAT_ID bit in
perf_event_attr.read_format so that when we read the event fds we get
that id to then hash it and be able later to use perf_evlist__id2evsel
to find the right evsel.

We also need to set the PERF_SAMPLE_ID bit in
perf_event_attr.sample_type to ask for that id to be stashed in each
sample, so that we can demux it.

So add a perf_evsel__set_sample_id() method to do those two things in
one operation.