top: Use perf_evlist__config()
Patchwork-id: 71342
O-Subject: [PATCH RHEL6.6 BZ1034843 036/172] perf top: Use perf_evlist__config()
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 2376c67a7bbc7849b806688ba2efb8520c21c458
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Dec 11 16:48:41 2012 -0300

description
===========
Using struct perf_record_opts to specify how to configure the evsel
perf_event_attrs.

This gets top closer to record in the way it sets up evsels, with the
aim of sharing more and more to the point that both will be a single
utility.

In this direction top now uses the same callchain option parsing as
record and that brings DWARF callchains to top, something that was
already available for record.