hists: Exchange order of comparing items when collapsing hists
Patchwork-id: 71325
O-Subject: [PATCH RHEL6.6 BZ1034843 037/172] perf hists: Exchange order of comparing items when collapsing hists
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 9afcf930b1fa1158b0878afeba3eff299300dc65
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Mon Dec 10 17:29:54 2012 +0900

description
===========
When comparing entries for collapsing put the given entry first, and
then the iterated entry.  This is not the case of hist_entry__cmp() when
called if given sort keys don't require collapsing.  So change the order
for the sake of consistency.  It will be required for matching and/or
linking multiple hist entries.