tests: Adjust some message log levels to help diagnosing problems in attr tests
Patchwork-id: 71344
O-Subject: [PATCH RHEL6.6 BZ1034843 042/172] perf tests: Adjust some message log levels to help diagnosing problems in attr tests
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit ce90e3856ba55af4d3d9c9c1168cc624607f6c7c
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Dec 18 14:14:24 2012 -0300

description
===========
Now we'll see the command being run and if it fails, the fields that had
unexpected values and the expected values, example testing a problem in the
next patch:

  # perf test -v 13
  13: struct perf_event_attr setup                           :
  --- start ---
  SNIP
    running 'PERF_TEST_ATTR=/tmp/tmpDNIE6M /home/acme/bin/perf record -o /tmp/tmpDNIE6M/perf.data --group -e cycles,instructions kill >/dev/null 2>&1' ret 0
    running 'PERF_TEST_ATTR=/tmp/tmpV5lKro /home/acme/bin/perf stat -o /tmp/tmpV5lKro/perf.data -dd kill >/dev/null 2>&1' ret 1
  expected config=3, got 65540
  expected exclude_guest=1, got 0
  FAILED '/home/acme/libexec/perf-core/tests/attr/test-stat-detailed-2' - match failure
  ---- end ----
  struct perf_event_attr setup: FAILED!
  #

While in the past we would see at the '-v' level many more messages for the
fields that matched, something we may want to see only in the '-vv' log level.

Keeping the 'running' messages so that we can see the tools tests that
succeeded so that we can compare it to the one that failed, helping pinpointing
the command line switch combo that leads to the problem.