tests: Add return states enum for tests
Patchwork-id: 71386
O-Subject: [PATCH RHEL6.6 BZ1034843 057/172] perf tests: Add return states enum for tests
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit f4c1ea5f2a6b9e1c0aaa874ffb25fe4a4f9f1a3f
Author: Jiri Olsa <jolsa@redhat.com>
Date: Wed Dec 19 11:33:39 2012 -0300

description
===========
Test can currently return one of 3 states: ok, fail, skip.

The ok and fail states are self-explanatory. The skip state means that
some of the conditions for running the test was not met, making it
impossible to even run the test. For instance, if the hardware doesn't
support the 'precise' level required by a test, it will be skipped.