probe: Allow of casting an array of char to string
Patchwork-id: 71371
O-Subject: [PATCH RHEL6.6 BZ1034843 069/172] perf probe: Allow of casting an array of char to string
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 7ce28b5b5b320e26ac6a0e352d5005bce3530e05
Author: Hyeoncheol Lee <hyc.lee@gmail.com>
Date: Tue Sep 11 16:57:28 2012 +0900

description
===========
Before casting a type of a variable to string, convert_variable_type()
confirms that the type is a pointer or an array. then if it is a pointer
to char, it is casted to string.  but in case of an array of char, it
isn't