sort: Fix --sort pid output
Patchwork-id: 71456
O-Subject: [PATCH RHEL6.6 BZ1034843 072/172] perf sort: Fix --sort pid output
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit fb29a338b585ebcce793b8e4a6c62440c4574fa7
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Thu Dec 27 18:11:40 2012 +0900

description
===========
The "pid" sort key prints "Command: Pid" output but it's misaligned.
It's because of the offset of 6 was added to the column length during
the calculation in order to reserve an space for Pid part but it isn't
honored when printed.  The output before this patch was like this:

  # Overhead  Command:  Pid      Shared Object
  # ........  .............  .................
  #
      99.70%        noploop:17814  noploop
       0.29%        noploop:17814  [kernel.kallsyms]
       0.01%        noploop:17814  ld-2.15.so

Fix it by subtracting 6 for printing comm part.