script: hook up perf_scripting_context->pevent
Patchwork-id: 71431
O-Subject: [PATCH RHEL6.6 BZ1034843 096/172] perf script: hook up perf_scripting_context->pevent
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 2de9533d6d61d3086a7079bf142d2bfa374e664e
Author: Tom Zanussi <tom.zanussi@linux.intel.com>
Date: Fri Jan 18 13:51:27 2013 -0600

description
===========
Running the check-perf-trace scripts causes segfaults in both the Perl
and Python cases:

  # perf script record check-perf-trace
  # perf script -s libexec/perf-core/scripts/python/check-perf-trace.py
  trace_begin
  Segmentation fault (core dumped)

The reason is that the 'pevent' field was added to
perf_scripting_context but it wasn't hooked up with an actual pevent in
either case, so when one of the 'common' fields is accessed (in
util/trace-event-parse.c:get_common_fields()), pevent->events tries to
dereference a NULL pointer.

This sets the pevent field when the scripting context is set up.