hists browser: Convert hpp helpers to a function
Patchwork-id: 71413
O-Subject: [PATCH RHEL6.6 BZ1034843 128/172] perf hists browser: Convert hpp helpers to a function
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 5aed9d24934be5b7fec1b66cc2a5f29fab4ec11e
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Tue Jan 22 18:09:35 2013 +0900

description
===========
The hpp helpers do same job for each field so it was implemented as
macro in order to access those fields easily.  But it gets cumbersome
to maintain a large function in a macro as the function grows. Factor
it out to a function with a little helper macro to access field.