gtk/browser: Convert hpp helpers to a function
Patchwork-id: 71415
O-Subject: [PATCH RHEL6.6 BZ1034843 129/172] perf gtk/browser: Convert hpp helpers to a function
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 843985e953ddcc3d57a62641b377c8d3222859e2
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Tue Jan 22 18:09:36 2013 +0900

description
===========
The hpp helpers do same job for each field so it was implemented as
macro in order to access those fields easily.  But it gets cumbersome to
maintain a large function in a macro as the function grows. Factor it
out to a function with a little helper macro to access field.