evlist: Fix set event list leader
Patchwork-id: 71437
O-Subject: [PATCH RHEL6.6 BZ1034843 142/172] perf evlist: Fix set event list leader
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 74b2133d19e776924b2773e27dd9d6940f1cc594
Author: Stephane Eranian <eranian@google.com>
Date: Thu Jan 31 13:54:37 2013 +0100

description
===========
The __perf_evlist__set_leader() was setting the leader for all events in
the list except the first. Which means it assumed the first event
already had event->leader = event.

Seems like this should be the role of the function to also do this. This
is a requirement for an upcoming patch set.