sort: Check return value of strdup()
Patchwork-id: 71459
O-Subject: [PATCH RHEL6.6 BZ1034843 146/172] perf sort: Check return value of strdup()
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 5936f54d6ca2857d81188dcdff8c61b8fc482f53
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Wed Feb 6 14:57:17 2013 +0900

description
===========
When setup_sorting() is called, 'str' is passed to strtok_r() but it's
not checked to have a valid pointer.  As strtok_r() accepts NULL pointer
on a first argument and use the third argument in that case, it can
cause a trouble since our third argument, tmp, is not initialized.