stat: Add per processor socket count aggregation
Patchwork-id: 71439
O-Subject: [PATCH RHEL6.6 BZ1034843 149/172] perf stat: Add per processor socket count aggregation
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit d7e7a451c13e784f497c054f1bd083d77be87498
Author: Stephane Eranian <eranian@google.com>
Date: Wed Feb 6 15:46:02 2013 +0100

description
===========
This patch adds per-processor socket count aggregation for system-wide
mode measurements. This is a useful mode to detect imbalance between
sockets.

To enable this mode, use --aggr-socket in addition
to -a. (system-wide).

The output includes the socket number and the number of online
processors on that socket. This is useful to gauge the amount of
aggregation.

 # ./perf stat -I 1000 -a --aggr-socket -e cycles sleep 2
 #           time socket cpus             counts events
      1.000097680 S0        4          5,788,785 cycles
      2.000379943 S0        4         27,361,546 cycles
      2.001167808 S0        4            818,275 cycles

[ committer note: Added missing man page entry based on above comments ]