annotate: Fix warning message on a missing vmlinux
Patchwork-id: 71421
O-Subject: [PATCH RHEL6.6 BZ1034843 157/172] perf annotate: Fix warning message on a missing vmlinux
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit e3a34029c635b7dee06e51d99441578b96c7d463
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Thu Feb 7 18:02:12 2013 +0900

description
===========
When perf annotate runs with no vmlinux file it cannot annotate kernel
symbols because the kallsyms only provides symbol addresses.  So it
recommends to run perf buildid-cache to install proper vmlinux image.

But running perf buildid-cache -av vmlinux as the message gives me a
following error:

  $ perf buildid-cache -av /home/namhyung/build/kernel/vmlinux
  Couldn't add v: No such file or directory

Since the -a option receives a parameter, 'v' should not be after the
option.

In addition -a option is not work for this case since the build-id cache
already has a kallsyms with same build-id so it'll fail with EEXIST.
Use recently added -u (--update) option for it.