gtk/annotate: Show source lines with gray color
Patchwork-id: 71465
O-Subject: [PATCH RHEL6.6 BZ1034843 160/172] perf gtk/annotate: Show source lines with gray color
Bugzilla: 1034843
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1034843
https://bugzilla.redhat.com/show_bug.cgi?id=1034843

upstream
========
commit 237522378604a2e26e19a8b11a70171eaf98c6c5
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Thu Feb 7 18:02:10 2013 +0900

description
===========
In order to differentiate source lines from asm line, print them with
gray color.  To do this, it needs to be escaped since sometimes it
contains "<" and/or ">" characters so that it should not be considered
as a markup tags.  Use glib's g_markup_escape_text() for this.